Updated payment status flag

+++ app/Http/Transformers/SittersTransformer.php
@@ -102,7 +102,7 @@ class SittersTransformer extends Transformer
                     'total'         => (float) $payment->total,
                     'tip'           => (float) $payment->tip,
                     'description'   => $payment->description,
-                     'payment_status'=> isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                     'payment_status'=> (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
                     'payment_via'=> $this->nulltoBlank($payment->payment_via),
                     'payment_details'=> $this->nulltoBlank($payment->payment_details)
                 ];
@@ -166,7 +166,7 @@ class SittersTransformer extends Transformer
                 "babies"            => [],
                 "payment"           => $paymentData,
                 'babies'            => $babyData,
-                'payment_status'    => isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                'payment_status'    => (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
             ];
             $sr++;
         }
@@ -201,7 +201,7 @@ class SittersTransformer extends Transformer
                     'total'         => (float) $payment->total,
                     'tip'           => (float) $payment->tip,
                     'description'   => $payment->description,
-                     'payment_status'=> isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                     'payment_status'=> (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
                     'payment_via'=> $this->nulltoBlank($payment->payment_via),
                     'payment_details'=> $this->nulltoBlank($payment->payment_details)
                 ];
@@ -265,7 +265,7 @@ class SittersTransformer extends Transformer
                 "babies"            => [],
                 "payment"           => $paymentData,
                 'babies'            => $babyData,
-                'payment_status'    => isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                'payment_status'    => (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
             ];
 
             $sr++;
@@ -365,7 +365,7 @@ class SittersTransformer extends Transformer
                 "babies"            => [],
                 "payment"           => $paymentData,
                 'babies'            => $babyData,
-                'payment_status'    => isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                'payment_status'    => (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
             ];
         }
 
@@ -404,7 +404,7 @@ class SittersTransformer extends Transformer
                     'total'         => (float) $payment->total,
                     'tip'           => (float) $payment->tip,
                     'description'   => $payment->description,
-                     'payment_status'=> isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                     'payment_status'=> (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
                     'payment_via'=> $this->nulltoBlank($payment->payment_via),
                     'payment_details'=> $this->nulltoBlank($payment->payment_details)
                 ];
@@ -468,7 +468,7 @@ class SittersTransformer extends Transformer
                 "babies"            => [],
                 "payment"           => $paymentData,
                 'babies'            => $babyData,
-                'payment_status'    => isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
+                'payment_status'    => (int) isset($payment->payment_status) ? $this->nulltoBlank($payment->payment_status) : 0,
             ];
 
             $sr++;
